refine, cod refine.

+++ consensus/tendermint/accountability/fault_detector.go
@@ -48,7 +48,7 @@ const (
 	offChainAccusationProofWindow = 10                           // the time window in block for one to provide off chain innocence proof before it is escalated on chain.
 	maxAccusationPerHeight        = 4                            // max number of accusation allowed to be produced by rule engine over a height against a validator.
 	maxNumOfInnocenceProofCached  = 120 * maxAccusationPerHeight // 120 blocks with 4 on each height that rule engine can produce totally over a height.
-	reportingSlotPeriod           = 20                           // Each AFD reporting slot holds 20 blocks, each validator response for a slot.
+	reportingSlotPeriod           = uint64(20)                   // Each AFD reporting slot holds 20 blocks, each validator response for a slot.
 )
 
 var (
